Automatic pwm steering
We would like the system to be as autonomous as it can be.
Autosteer provided in this commit checks if the temperature during last 15
seconds has stabilized, and if it stabilized, pwm is driven higher or lower
depending on desired temperature. 15 second time period has been chosen
based on experience in furnace heating.