Remove workaround for PWM_MAX smoothing
Pico's PWM driver has this weird behavior of going down to 0V when close
to wrapping. Earlier we had a *very smart* way for dealing with it in
software, but it turns out It doesn't always work and we observed it
going down to 0V sometimes.

For now we don't care about the small drop at the end. It's not ideal,
but It's better than not working code.