Introduce saving user config to flash memory
Now, device is saving user config to non volatile memory.
We are declaring space for user config in the internal device's
flash memory.

The main purpose of this utility, is to handle unexpected
loss of electrical power.
We want to let the device to boot itself up with previous config.