BUGFIX: CONFIG_THERMO not works
The GNU standard says:
"Identifiers that are not macros, which are all considered to be the
number zero. This allows you to write #if MACRO instead of #ifdef MACRO,
if you know that MACRO, when defined, will always have a nonzero value.
Function-like macros used without their function call parentheses are
also treated as zero."

According to that, pt100 and ktype were identifiers that are not macros,
and was considered as 0.
So we were always including, no matter what the user has passed.