pwm: lower the frequency to 8Hz
We confirmed empirically that using "high" frequency as 40Hz is causing
the resulting signal to be a complete nonsense. In practice we were
seeing a long windows of even 5.5 seconds of constant, full power 230V
sine followed by a similar, long windows of 0V. This behavior was
causing the resistance wire to break frequently, making the furnace
unusable. While in theory 8Hz should result in less smooth heating, we
confirmed empirically, that in fact it's much more reliable.

The whole power delivery circuitry needs to be and is being redesigned,
but for now we ship a fairly easy work-around which allows us to use the
furnace, while we wait for a better power delivery system to be built.